requirements.txt
Pillow==9.2.0
watchdog==2.1.9
